Refactor backend Rust modules to reduce file size
Split oversized backend modules, deduplicate provider response handling, and trim repeated command and pipeline boilerplate without changing command contracts or credential behavior.
