+++ src-tauri/src/data/db/schema.rs
@@ -186,16 +186,15 @@ pub fn open(path: impl AsRef<std::path::Path>) -> Result<Db> {
         }
     }
     if user_version < 4 {
-        conn.execute_batch("BEGIN;")?;
-        if let Err(err) = (|| -> Result<()> {
+        run_migration(&conn, |conn| {
             ensure_table_column(
-                &conn,
+                conn,
                 "dictionary",
                 "confidence_tier",
                 "ALTER TABLE dictionary ADD COLUMN confidence_tier TEXT NOT NULL DEFAULT 'low';",
             )?;
             ensure_table_column(
-                &conn,
+                conn,
                 "dictionary",
                 "last_seen_at",
                 "ALTER TABLE dictionary ADD COLUMN last_seen_at DATETIME;",
@@ -230,53 +229,34 @@ pub fn open(path: impl AsRef<std::path::Path>) -> Result<Db> {
                  PRAGMA user_version = 4;",
             )?;
             Ok(())
-        })() {
-            let _ = conn.execute_batch("ROLLBACK;");
-            return Err(err);
-        }
-        conn.execute_batch("COMMIT;")?;
+        })?;
     }
     if user_version < 5 {
-        conn.execute_batch("BEGIN;")?;
-        if let Err(err) = (|| -> Result<()> {
-            ensure_cleanup_cache_schema(&conn)?;
+        run_migration(&conn, |conn| {
+            ensure_cleanup_cache_schema(conn)?;
             conn.execute_batch("PRAGMA user_version = 5;")?;
             Ok(())
-        })() {
-            let _ = conn.execute_batch("ROLLBACK;");
-            return Err(err);
-        }
-        conn.execute_batch("COMMIT;")?;
+        })?;
     }
     if user_version < 6 {
-        conn.execute_batch("BEGIN;")?;
-        if let Err(err) = (|| -> Result<()> {
-            ensure_cleanup_cache_schema(&conn)?;
+        run_migration(&conn, |conn| {
+            ensure_cleanup_cache_schema(conn)?;
             conn.execute_batch("PRAGMA user_version = 6;")?;
             Ok(())
-        })() {
-            let _ = conn.execute_batch("ROLLBACK;");
-            return Err(err);
-        }
-        conn.execute_batch("COMMIT;")?;
+        })?;
     }
     if user_version < 7 {
-        conn.execute_batch("BEGIN;")?;
-        if let Err(err) = (|| -> Result<()> {
+        run_migration(&conn, |conn| {
             ensure_table_column(
-                &conn,
+                conn,
                 "transcriptions",
                 "spoken_words",
                 "ALTER TABLE transcriptions ADD COLUMN spoken_words INTEGER;",
             )?;
-            backfill_spoken_words(&conn)?;
+            backfill_spoken_words(conn)?;
             conn.execute_batch("PRAGMA user_version = 7;")?;
             Ok(())
-        })() {
-            let _ = conn.execute_batch("ROLLBACK;");
-            return Err(err);
-        }
-        conn.execute_batch("COMMIT;")?;
+        })?;
     }
     if user_version < 8 {
         let res = conn.execute_batch(
@@ -336,6 +316,16 @@ pub fn open(path: impl AsRef<std::path::Path>) -> Result<Db> {
     }
 
     Ok(Arc::new(Mutex::new(conn)))
+}
+
+fn run_migration(conn: &Connection, f: impl FnOnce(&Connection) -> Result<()>) -> Result<()> {
+    conn.execute_batch("BEGIN;")?;
+    if let Err(err) = f(conn) {
+        let _ = conn.execute_batch("ROLLBACK;");
+        return Err(err);
+    }
+    conn.execute_batch("COMMIT;")?;
+    Ok(())
 }
 
 pub fn table_has_column(conn: &Connection, table: &str, column: &str) -> Result<bool> {
